+++ resources/views/components/datasets/org_datasets.blade.php
@@ -1,28 +1,21 @@
-@extends('layouts.app')
-
-@section('content')
-<div class="container">
-    @include('partials.alerts-bar')
-    @include('partials.user-nav-bar', ['view' => 'dataset'])
-    <div class="col-xs-12 sidenav m-t-lg m-b-lg">
-        <span class="my-profile m-l-sm">{{uctrans('custom.datasets_list')}}</span>
-    </div>
-    @include('partials.pagination')
-    <div class="row">
-    @if ($buttons['add'])
+@php $root = empty($admin) ? 'user' : 'admin'; @endphp
+<div class="row">
+    @if ((!empty($admin)) || !empty($buttons['add']))
         <div class="col-sm-3 col-xs-12 text-left">
             <span class="badge badge-pill m-t-lg new-data user-add-btn">
-                <a href="{{ url('/user/dataset/create') }}">{{ __('custom.add_new_dataset') }}</a>
+                <a
+                    href="{{ url('/'. $root .'/organisations/dataset/create/'. $organisation->uri) }}"
+                >{{ __('custom.add_new_dataset') }}</a>
             </span>
         </div>
     @endif
-    @if ($buttons['view'])
+    @if ((!empty($admin)) || !empty($buttons['view']))
         <div class="col-lg-8 col-md-7 col-sm-12 col-xs-12 search-field">
             <form method="GET">
                 <input
                     type="text"
                     class="m-t-md input-border-r-12 form-control js-ga-event"
-                    placeholder="{{ __('custom.search') }}.."
+                    placeholder="{{ __('custom.search') }}"
                     value="{{ isset($search) ? $search : '' }}"
                     name="q"
                     data-ga-action="search"
@@ -32,43 +25,38 @@
             </form>
         </div>
     @endif
-    </div>
-    <div class="row">
-        <div class="col-xs-12 m-t-md">
+</div>
+<div class="row">
+    @include('partials.org-info', ['organisation' => $organisation])
+    <div class="col-sm-9 col-xs-12 m-t-md">
+        <div class="row">
             <div class="articles m-t-lg">
                 @if (count($datasets))
                     @foreach ($datasets as $set)
-                        @if ($buttons[$set->uri]['view'])
                         <div class="article m-b-lg col-xs-12 user-dataset">
                             <div>{{ __('custom.date_added') }}: {{ $set->created_at }}</div>
                             <div class="col-sm-12 p-l-none">
-                                <a href="{{ route('datasetView', ['uri' => $set->uri]) }}">
-                                    <h2 class="m-t-xs {{$set->reported ? 'error' : '' }}">{{ $set->name }}</h2>
+                                <a href="{{ url('/'. $root .'/organisations/dataset/view/'. $set->uri) }}">
+                                    <h2 class="m-t-xs">{{ $set->name }}</h2>
                                 </a>
-                                @if ($set->status == 1)
-                                    <span>({{ __('custom.draft') }})</span>
-                                @else
-                                    <span>({{ __('custom.published') }})</span>
-                                @endif
                                 <div class="desc">
-                                    {!! nl2br(e($set->descript)) !!}
+                                    {!! nl2br(truncate(e($set->descript), 150)) !!}
                                 </div>
                                 <div class="col-sm-12 p-l-none btns">
                                     <div class="pull-left row">
+                                    @if ((!empty($admin)) || (!empty($buttons[$set->uri]['edit'])))
                                         <div class="col-xs-6">
-                                        @if ($buttons[$set->uri]['edit'])
                                             <span class="badge badge-pill m-r-md m-b-sm">
-                                                <a href="{{ url('/user/dataset/edit/'. $set->uri) }}">
-                                                    {{ uctrans('custom.edit') }}
-                                                </a>
+                                                <a
+                                                    href="{{ url('/'. $root .'/organisation/'. $organisation->uri .'/dataset/edit/'. $set->uri) }}"
+                                                >{{ uctrans('custom.edit') }}</a>
                                             </span>
-                                        @endif
                                         </div>
-                                        <div class="col-xs-6">
-                                        @if ($buttons[$set->uri]['delete'])
-                                            <form method="POST">
-                                                {{ csrf_field() }}
-                                                <input type="hidden" name="dataset_uri" value="{{ $set->uri }}">
+                                    @endif
+                                    <div class="col-xs-6">
+                                        <form method="POST">
+                                            {{ csrf_field() }}
+                                            @if ((!empty($admin)) || (!empty($buttons[$set->uri]['delete'])))
                                                 <div class="col-xs-6 text-right">
                                                     <button
                                                         class="badge badge-pill m-b-sm del-btn"
@@ -77,21 +65,21 @@
                                                         data-confirm="{{ __('custom.remove_data') }}"
                                                     >{{ uctrans('custom.remove') }}</button>
                                                 </div>
-                                            </form>
                                             @endif
-                                        </div>
+                                            <input type="hidden" name="dataset_uri" value="{{ $set->uri }}">
+                                        </form>
+                                    </div>
                                     </div>
                                     <div class="pull-right">
                                         <span>
-                                            <a href="{{ route('datasetView', ['uri' => $set->uri]) }}">
-                                                {{ __('custom.see_more') }}
-                                            </a>
+                                            <a
+                                                href="{{ url('/'. $root .'/organisations/dataset/view/'. $set->uri) }}"
+                                            >{{ __('custom.see_more') }}</a>
                                         </span>
                                     </div>
                                 </div>
                             </div>
                         </div>
-                        @endif
                     @endforeach
                 @else
                     <div class="col-sm-12 m-t-xl text-center no-info">
@@ -101,12 +89,4 @@
             </div>
         </div>
     </div>
-    @if (isset($pagination))
-        <div class="row">
-            <div class="col-xs-12 text-center pagination">
-                {{ $pagination->render() }}
-            </div>
-        </div>
-    @endif
 </div>
-@endsection