+++ resources/views/user/resourceView.blade.php
@@ -4,6 +4,8 @@
 <div class="container">
     @include('partials.alerts-bar')
     @include('partials.user-nav-bar', ['view' => 'user'])
-    @include('components.datasets.resource_view')
+    <div class="col-xs-12 m-t-md">
+        @include('components.datasets.resource_view')
+    </div>
 </div>
 @endsection